+++ src/components/forms/ScmToolRadioGroup.js
@@ -2,7 +2,6 @@ import React from 'react';
 
 import Radio from './Radio';
 import HelpText from './HelpText';
-import { INPUT_COLORS } from '.';
 
 export const SCM_TOOLS = [{
   value: 'github-cloud',
@@ -59,43 +58,5 @@ const ScmToolRadioGroup = ({
     </div>
   </fieldset>
 );
-
-export const ScmToolSelect = ({
-  onChange,
-  currentValue,
-  label = 'Primary source code host',
-  idPrefix = '',
-  color = 'primary',
-}) => {
-  const { accent, border, placeholder, background, text, label: labelStyle } = INPUT_COLORS[color];
-  const onInputChange = (e) => {
-    onChange(e.target.value);
-  };
-  return (
-    <>
-      <label className={`block text-sm font-medium ${labelStyle}`} htmlFor="scm-select">
-        {label}
-      </label>
-      <select
-        className={`block mt-1.5 appearance-none w-full py-3 px-4 block shadow-sm rounded-md ${background} ${text} ${accent} ${border} ${placeholder}`}
-        id="scm-select"
-        value={currentValue}
-        name="scm-select"
-        onChange={onInputChange}
-      >
-        {SCM_TOOLS.map(({ value, label }) => (
-          <option
-            key={`sct-option-${value}`}
-            value={value}
-            name={`sct-option-${value}`}
-            id={`${idPrefix}scm-${value}-input`}
-          >
-            {label}
-          </option>
-        ))}
-      </select>
-    </>
-  );
-};
 
 export default ScmToolRadioGroup;
